requirements.txt
streamlit>=1.0
torch>=1.12
timm>=0.6.12
torchvision
Pillow
numpy
opencv-python-headless
scikit-learn
matplotlib
